+++ packages/storage/src/bucket.rs
@@ -67,14 +67,14 @@ where
 
     /// load will return an error if no data is set at the given key, or on parse error
     pub fn load(&self, key: &[u8]) -> StdResult<T> {
-        let value = get_with_prefix(self.storage, &self.prefix, key)?;
+        let value = get_with_prefix(self.storage, &self.prefix, key);
         must_deserialize(&value)
     }
 
     /// may_load will parse the data stored at the key if present, returns Ok(None) if no data there.
     /// returns an error on issues parsing
     pub fn may_load(&self, key: &[u8]) -> StdResult<Option<T>> {
-        let value = get_with_prefix(self.storage, &self.prefix, key)?;
+        let value = get_with_prefix(self.storage, &self.prefix, key);
         may_deserialize(&value)
     }
 
@@ -139,14 +139,14 @@ where
 
     /// load will return an error if no data is set at the given key, or on parse error
     pub fn load(&self, key: &[u8]) -> StdResult<T> {
-        let value = get_with_prefix(self.storage, &self.prefix, key)?;
+        let value = get_with_prefix(self.storage, &self.prefix, key);
         must_deserialize(&value)
     }
 
     /// may_load will parse the data stored at the key if present, returns Ok(None) if no data there.
     /// returns an error on issues parsing
     pub fn may_load(&self, key: &[u8]) -> StdResult<Option<T>> {
-        let value = get_with_prefix(self.storage, &self.prefix, key)?;
+        let value = get_with_prefix(self.storage, &self.prefix, key);
         may_deserialize(&value)
     }